refactor(Ranking): 랭킹 관련 컴포넌트 modals 폴더로 이동
